requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aggdraw==1.3.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
certifi==2021.5.30
charset-normalizer==2.0.4; python_version >= '3'
cloudinary==1.26.0
colorama==0.4.4; platform_system == 'Windows'
contextlib2==21.6.0; python_version >= '3.6'
cycler==0.10.0
docopt==0.6.2
idna==3.2; python_version >= '3'
imageio==2.9.0
kiwisolver==1.3.1; python_version >= '3.6'
lxml==4.6.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
matplotlib==3.4.2; python_version >= '3.7'
networkx==2.6.2; python_version >= '3.7'
numpy==1.21.1; python_version >= '3.7'
pathlib2==2.3.6
pillow==8.3.1; python_version >= '3.6'
psd-tools==1.9.17
pyaes==1.6.1
pyasn1==0.4.8
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pysocks==1.7.1
pytelegrambotapi==3.8.2
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pywavelets==1.1.1; python_version >= '3.5'
pyyaml==5.4.1
requests==2.26.0
rsa==4.7.2; python_version >= '3.5' and python_version < '4'
scikit-image==0.18.2; python_version >= '3.7'
scipy==1.7.0; python_version < '3.10' and python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
svgutils==0.3.4
telethon==1.23.0
temporary==3.0.0
tifffile==2021.7.30; python_version >= '3.7'
tqdm==4.62.0
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
